Combat Systems and Loot (ch15)

+++ src/map.rs
@@ -92,6 +92,6 @@ impl BaseMap for Map {
     }
 
     fn is_opaque(&self, idx: usize) -> bool {
-        self.tiles[idx as usize] != TileType::Floor
+        self.tiles[idx] != TileType::Floor
     }
 }
